# api/openapi.yaml
openapi: 3.1.0
info:
  title: Eval Hub
  description: API REST server for evaluation backend orchestration on OpenShift
  version: 0.1.0
  contact:
    name: API Support
servers:
- url: http://localhost:8080
  description: Local development server
- url: https://api.example.com
  description: Production server
paths:
...
  /api/v1/status:
    get:
      summary: Get service status
      description: Returns detailed status information about the service
      operationId: getStatus
      tags:
      - Status
      responses:
        '200':
          description: Service status information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '405':
          description: Method not allowed
          content:
            text/plain:
              schema:
                type: string
                example: Method not allowed
...
  /api/v1/evaluations/jobs:
    post:
      tags:
      - Evaluations
      summary: Create Evaluation
      description: Create and execute evaluation request using the simplified benchmark schema.
      operationId: create_evaluation_api_v1_evaluations_jobs_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimpleEvaluationRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Evaluations
      summary: List Evaluations
      description: List all evaluation requests.
      operationId: list_evaluations_api_v1_evaluations_jobs_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Maximum number of evaluations to return
          default: 50
          title: Limit
        description: Maximum number of evaluations to return
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Offset for pagination
          default: 0
          title: Offset
        description: Offset for pagination
      - name: status_filter
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by status
          title: Status Filter
        description: Filter by status
      - name: summary
        in: query
        required: false
        schema:
          type: boolean
          description: If true, return summary view without detailed results
          default: false
          title: Summary
        description: If true, return summary view without detailed results
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEvaluations'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/evaluations/benchmarks:
    get:
      tags:
      - Benchmarks
      summary: List All Benchmarks
      description: List all available benchmarks across providers (similar to Llama Stack format).
      operationId: list_all_benchmarks_api_v1_evaluations_benchmarks_get
      parameters:
      - name: provider_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by provider ID
          title: Provider Id
        description: Filter by provider ID
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by benchmark category
          title: Category
        description: Filter by benchmark category
      - name: tags
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by tags (comma-separated)
          title: Tags
        description: Filter by tags (comma-separated)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListBenchmarksResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/evaluations/collections:
    get:
      tags:
      - Collections
      summary: List Collections
      description: List all benchmark collections.
      operationId: list_collections_api_v1_evaluations_collections_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListCollectionsResponse'
    post:
      tags:
      - Collections
      summary: Create Collection
      description: Create a new collection.
      operationId: create_collection_api_v1_evaluations_collections_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreationRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    StatusResponse:
      type: object
      required:
      - service
      - version
      - status
      - timestamp
      properties:
        service:
          type: string
          example: eval-hub-backend-svc
          description: Name of the service
        version:
          type: string
          example: 1.0.0
          description: Version of the service
        status:
          type: string
          example: running
          description: Current status of the service
        timestamp:
          type: string
          format: date-time
          example: '2024-01-13T10:23:48Z'
          description: Timestamp of the status check in RFC3339 format
...
    BenchmarkConfig:
      properties:
        benchmark_id:
          type: string
          title: Benchmark Id
          description: Benchmark identifier
        provider_id:
          type: string
          title: Provider Id
          description: Provider identifier
        config:
          additionalProperties: true
          type: object
          title: Config
          description: Benchmark configuration including num_fewshot, limit, batch_size, etc.
      additionalProperties: false
      type: object
      required:
      - benchmark_id
      - provider_id
      title: BenchmarkConfig
      description: New simplified benchmark specification.
    BenchmarkReference:
      properties:
        provider_id:
          type: string
          title: Provider Id
          description: Provider identifier
        benchmark_id:
          type: string
          title: Benchmark Id
          description: Benchmark identifier
        weight:
          type: number
          title: Weight
          description: Weight for this benchmark in collection scoring
          default: 1.0
        config:
          additionalProperties: true
          type: object
          title: Config
          description: Benchmark-specific configuration
      additionalProperties: true
      type: object
      required:
      - provider_id
      - benchmark_id
      title: BenchmarkReference
      description: Reference to a benchmark within a collection.
    BenchmarkResultPayload:
      properties:
        name:
          type: string
          title: Name
          description: Benchmark name
        metrics:
          additionalProperties:
            anyOf:
            - type: number
            - type: integer
            - type: string
          type: object
          title: Metrics
          description: Benchmark metrics
        artifacts:
          additionalProperties:
            type: string
          type: object
          title: Artifacts
          description: Benchmark artifacts
        mlflow_run_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Mlflow Run Id
          description: MLFlow run ID
      additionalProperties: true
      type: object
      required:
      - name
      title: BenchmarkResultPayload
      description: Result payload for a single benchmark.
    Collection:
      properties:
        collection_id:
          type: string
          title: Collection Id
          description: Unique collection identifier
        name:
          type: string
          title: Name
          description: Human-readable collection name
        description:
          type: string
          title: Description
          description: Collection description
        provider_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider Id
          description: Primary provider for this collection
        tags:
          items:
            type: string
          type: array
          title: Tags
          description: Tags for categorizing the collection
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
          description: Additional collection metadata
        benchmarks:
          items:
            $ref: '#/components/schemas/BenchmarkReference'
          type: array
          title: Benchmarks
          description: List of benchmark references in this collection
        created_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Created At
          description: Collection creation timestamp
        updated_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Updated At
          description: Collection last update timestamp
      additionalProperties: true
      type: object
      required:
      - collection_id
      - name
      - description
      - benchmarks
      title: Collection
      description: Collection of benchmarks for specific evaluation scenarios.
    CollectionCreationRequest:
      properties:
        collection_id:
          type: string
          title: Collection Id
          description: Unique collection identifier
        name:
          type: string
          title: Name
          description: Human-readable collection name
        description:
          type: string
          title: Description
          description: Collection description
        provider_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider Id
          description: Primary provider for this collection
        tags:
          items:
            type: string
          type: array
          title: Tags
          description: Tags for categorizing the collection
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
          description: Additional collection metadata
        benchmarks:
          items:
            $ref: '#/components/schemas/BenchmarkReference'
          type: array
          title: Benchmarks
          description: List of benchmark references in this collection
      additionalProperties: true
      type: object
      required:
      - collection_id
      - name
      - description
      - benchmarks
      title: CollectionCreationRequest
      description: Request for creating a new collection.
...
    EvaluationResponse:
      properties:
        system:
          $ref: '#/components/schemas/SystemInfo'
          description: System metadata for the request
        results:
          anyOf:
          - $ref: '#/components/schemas/ResultsPayload'
          - type: 'null'
          description: Results payload available upon success
        model:
          $ref: '#/components/schemas/Model'
          description: Model configuration provided by the user
        benchmarks:
          items:
            $ref: '#/components/schemas/BenchmarkConfig'
          type: array
          title: Benchmarks
          description: Benchmarks provided in the request
        experiment:
          $ref: '#/components/schemas/ExperimentConfig'
          description: Experiment configuration provided by the user
        timeout_minutes:
          type: integer
          title: Timeout Minutes
          description: Timeout for the entire evaluation (user-provided)
        retry_attempts:
          type: integer
          title: Retry Attempts
          description: Number of retry attempts (user-provided)
        callback_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Callback Url
          description: Callback URL provided by the user
        async_mode:
          type: boolean
          title: Async Mode
          description: Whether the evaluation runs asynchronously
        custom:
          additionalProperties: true
          type: object
          title: Custom
          description: Custom fields supplied by the user in the request
      additionalProperties: true
      type: object
      required:
      - system
      - model
      - benchmarks
      - experiment
      - timeout_minutes
      - retry_attempts
      - async_mode
      title: EvaluationResponse
      description: Response payload for evaluation requests.
...
    ExperimentConfig:
      properties:
        name:
          type: string
          title: Name
          description: Experiment name for MLFlow tracking
        tags:
          additionalProperties:
            type: string
          type: object
          title: Tags
          description: Tags for the evaluation experiment
      additionalProperties: false
      type: object
      required:
      - name
      title: ExperimentConfig
      description: Configuration for MLFlow experiment tracking.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ListBenchmarksResponse:
      properties:
        benchmarks:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Benchmarks
          description: List of all available benchmarks
        total_count:
          type: integer
          title: Total Count
          description: Total number of benchmarks
        providers_included:
          items:
            type: string
          type: array
          title: Providers Included
          description: List of provider IDs included in the response
      additionalProperties: true
      type: object
      required:
      - benchmarks
      - total_count
      - providers_included
      title: ListBenchmarksResponse
      description: Response for listing all benchmarks (similar to Llama Stack format).
    ListCollectionsResponse:
      properties:
        collections:
          items:
            $ref: '#/components/schemas/Collection'
          type: array
          title: Collections
          description: List of available collections
        total_collections:
          type: integer
          title: Total Collections
          description: Total number of collections
      additionalProperties: true
      type: object
      required:
      - collections
      - total_collections
      title: ListCollectionsResponse
      description: Response for listing all collections.
...
    Model:
      properties:
        url:
          type: string
          title: Url
          description: Model endpoint URL
        name:
          type: string
          title: Name
          description: Model name/identifier
      additionalProperties: false
      type: object
      required:
      - url
      - name
      title: Model
      description: Model specification for evaluation requests.
    PaginatedEvaluations:
      properties:
        first:
          $ref: '#/components/schemas/PaginationLink'
          description: Link to the first page
        next:
          anyOf:
          - $ref: '#/components/schemas/PaginationLink'
          - type: 'null'
          description: Link to the next page, if available
        limit:
          type: integer
          title: Limit
          description: Page size used for this response
        total_count:
          type: integer
          title: Total Count
          description: Total number of evaluations
        items:
          items:
            $ref: '#/components/schemas/EvaluationResponse'
          type: array
          title: Items
          description: Evaluations returned for this page
      additionalProperties: true
      type: object
      required:
      - first
      - limit
      - total_count
      title: PaginatedEvaluations
      description: Paginated list response for evaluation resources.
    PaginationLink:
      properties:
        href:
          type: string
          title: Href
          description: Link URL
      additionalProperties: true
      type: object
      required:
      - href
      title: PaginationLink
      description: Hypermedia link used for pagination.
...
    ResultsPayload:
      properties:
        benchmarks:
          items:
            $ref: '#/components/schemas/BenchmarkResultPayload'
          type: array
          title: Benchmarks
          description: List of benchmark results
        aggregated_metrics:
          additionalProperties:
            anyOf:
            - type: number
            - type: integer
            - type: string
          type: object
          title: Aggregated Metrics
          description: Aggregated metrics across benchmarks
        mlflow_experiment_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Mlflow Experiment Url
          description: Link to the MLFlow experiment
      additionalProperties: true
      type: object
      title: ResultsPayload
      description: Container for evaluation results.
    RunStatus:
      properties:
        name:
          type: string
          title: Name
          description: User-provided benchmark name
        state:
          type: string
          title: State
          description: Run state (pending, running, success, failed)
        message:
          type: string
          title: Message
          description: Optional status message
          default: ''
        logs_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Logs Path
          description: Optional path to run-specific logs
      additionalProperties: true
      type: object
      required:
      - name
      - state
      title: RunStatus
      description: Status information for an individual benchmark run.
    SimpleEvaluationRequest:
      properties:
        model:
          $ref: '#/components/schemas/Model'
          description: Model specification for evaluation
        benchmarks:
          items:
            $ref: '#/components/schemas/BenchmarkConfig'
          type: array
          title: Benchmarks
          description: List of benchmarks to evaluate
        experiment:
          $ref: '#/components/schemas/ExperimentConfig'
          description: Experiment configuration for MLFlow tracking
        timeout_minutes:
          type: integer
          title: Timeout Minutes
          description: Timeout for the entire evaluation
          default: 60
        retry_attempts:
          type: integer
          title: Retry Attempts
          description: Number of retry attempts on failure
          default: 3
        async_mode:
          type: boolean
          title: Async Mode
          description: Whether to run evaluations asynchronously
          default: true
        callback_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Callback Url
          description: URL to call when evaluation completes
        created_at:
          type: string
          format: date-time
          title: Created At
          description: Request creation timestamp
        custom:
          additionalProperties: true
          type: object
          title: Custom
          description: Custom fields for additional request data
      additionalProperties: false
      type: object
      required:
      - model
      - benchmarks
      - experiment
      title: SimpleEvaluationRequest
      description: Simplified evaluation request using the new schema.
    SystemInfo:
      properties:
        id:
          type: string
          format: uuid
          title: Id
          description: Unique evaluation request ID
        status:
          $ref: '#/components/schemas/SystemStatus'
          description: Current status of the request
        created_at:
          type: string
          format: date-time
          title: Created At
          description: Creation timestamp
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
          description: Completion timestamp
      additionalProperties: true
      type: object
      required:
      - id
      - status
      - created_at
      title: SystemInfo
      description: Metadata about the evaluation resource.
    SystemStatus:
      properties:
        state:
          type: string
          title: State
          description: Overall state of the request
        message:
          type: string
          title: Message
          description: Optional status message
          default: ''
        logs_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Logs Path
          description: Optional path to system logs
        runs:
          items:
            $ref: '#/components/schemas/RunStatus'
          type: array
          title: Runs
          description: Status for each benchmark run
      additionalProperties: true
      type: object
      required:
      - state
      title: SystemStatus
      description: Overall status for an evaluation request.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...